Move all detoured function to harmony
