+++ Patch/CargoTruckAISetTargetPatch.cs
@@ -17,7 +17,7 @@ namespace RealGasStation.Patch
         {
             return typeof(CargoTruckAI).GetMethod("SetTarget", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Vehicle).MakeByRefType(), typeof(ushort) }, null);
         }
-        public static bool Prefix(ushort vehicleID, ref Vehicle data, ushort targetBuilding)
+        public static bool Prefix(ref CargoTruckAI __instance, ushort vehicleID, ref Vehicle data, ushort targetBuilding)
         {
             if ((data.m_transferType == 113) || (data.m_transferType == 112))
             {
@@ -42,8 +42,7 @@ namespace RealGasStation.Patch
                             instance.ReleasePath(data.m_path);
                             data.m_path = 0;
                         }
-                        CargoTruckAI AI = (CargoTruckAI)data.Info.m_vehicleAI;
-                        AI.SetTarget(vehicleID, ref data, data.m_targetBuilding);
+                        __instance.SetTarget(vehicleID, ref data, data.m_targetBuilding);
                         MainDataStore.TargetGasBuilding[vehicleID] = 0;
                     }
                     return false;
